feature: adding sidebar
